refac(setup): combine into data organization
instead of managing all these different steps in different files,
also to use similar subfunctions
